Finally refactor SmallButtonStyle as standalone component
